+++ Blockzilla/Assets.xcassets/icon_shortcuts.imageset/icon_shortcuts.pdf
@@ -1,0 +1,103 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 2.500244 2.425735 cm
+0.984314 0.984314 0.996078 scn
+18.779751 1.354263 m
+13.307750 6.826262 l
+16.293751 9.812263 l
+16.820751 10.340262 16.973749 11.097261 16.691750 11.788261 c
+16.409750 12.478262 15.771750 12.912262 15.026751 12.921263 c
+12.551750 12.950262 l
+8.776751 16.726263 l
+9.844750 17.794262 l
+10.137751 18.087263 10.137751 18.562262 9.844750 18.855263 c
+9.551750 19.148262 9.076751 19.148262 8.783751 18.855263 c
+0.219750 10.287263 l
+-0.073250 9.994263 -0.073250 9.519263 0.219750 9.226263 c
+0.365750 9.080263 0.557750 9.006263 0.749750 9.006263 c
+0.941750 9.006263 1.133750 9.079263 1.279750 9.226263 c
+2.358750 10.305263 l
+6.133750 6.530263 l
+6.162750 4.055264 l
+6.171750 3.309263 6.606750 2.671263 7.297750 2.390263 c
+7.526751 2.296263 7.762750 2.251263 7.994750 2.251263 c
+8.464750 2.251263 8.918750 2.436264 9.270750 2.788263 c
+12.247750 5.765263 l
+17.719749 0.293262 l
+17.865749 0.147263 18.057751 0.073263 18.249750 0.073263 c
+18.441750 0.073263 18.633751 0.146263 18.779751 0.293262 c
+19.072750 0.587263 19.072750 1.061264 18.779751 1.354263 c
+h
+8.210751 3.849262 m
+8.078751 3.717262 7.936750 3.749263 7.862750 3.779263 c
+7.789750 3.809262 7.664750 3.887263 7.662750 4.073263 c
+7.625751 7.159263 l
+3.418750 11.366262 l
+7.717751 15.665262 l
+11.923750 11.458261 l
+15.009750 11.421263 l
+15.195750 11.419262 15.273751 11.295261 15.303751 11.222261 c
+15.332751 11.148262 15.364749 11.004262 15.233749 10.873262 c
+8.210751 3.849262 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  1521
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 24.000000 24.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Type /Catalog
+     /Pages 5 0 R
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000001611 00000 n
+0000001634 00000 n
+0000001807 00000 n
+0000001881 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+1940
+%%EOF
